+++ back/engines/commercial/customizable_navbar/lib/customizable_navbar.rb
@@ -1,0 +1,4 @@
+require 'customizable_navbar/engine'
+
+module CustomizableNavbar
+end