+++ back/engines/commercial/nlp/app/controllers/nlp/web_api/v1/similar_ideas_controller.rb
@@ -2,9 +2,9 @@ module NLP
   module WebApi
     module V1
       class SimilarIdeasController < ApplicationController
+        skip_before_action :authenticate_user
 
         SIMILARITY_TRESHOLD = 0.15
-
 
         def index
           idea = Idea.find params[:idea_id]
@@ -13,23 +13,14 @@ module NLP
           service = NLP::SimilarityService.new
           similarities = service.similarity(
             Tenant.current.id, idea,
-            idea_ids: policy_scope(Idea).ids, 
+            idea_ids: policy_scope(Idea).ids,
             min_score: SIMILARITY_TRESHOLD
-            )
-          @ideas = policy_scope(Idea.where(id: similarities.map{|h| h[:idea_id]}))
-            .page(params.dig(:page, :number))
-            .per(params.dig(:page, :size))
+          )
+          @ideas = policy_scope(Idea.where(id: similarities.map{ |h| h[:idea_id] }))
+          @ideas = paginate @ideas
 
           render json: linked_json(@ideas, ::WebApi::V1::SimilarIdeaSerializer, params: fastjson_params)
         end
-
-
-        private
-
-        def secure_controller?
-          false
-        end
-
       end
     end
   end