+++ back/engines/free/email_campaigns/app/controllers/email_campaigns/web_api/v1/consents_controller.rb
@@ -1,16 +1,14 @@
 module EmailCampaigns
   class WebApi::V1::ConsentsController < EmailCampaignsController
-
-    before_action :set_consent, only: [:update]
-    before_action :ensure_consents, only: [:index, :update, :update_by_campaign_id]
+    before_action :set_consent, only: :update
+    before_action :ensure_consents, only: %i[index update update_by_campaign_id]
+    skip_before_action :authenticate_user
 
     def index
       authorize Consent
-      
-      @consents = policy_scope(Consent)
-        .where(user: current_user_by_unsubscription_token)
-        .page(params.dig(:page, :number))
-        .per(params.dig(:page, :size))
+
+      @consents = policy_scope(Consent).where(user: current_user_by_unsubscription_token)
+      @consents = paginate @consents
 
       render json: linked_json(@consents, WebApi::V1::ConsentSerializer, params: fastjson_params)
     end
@@ -34,7 +32,6 @@ module EmailCampaigns
       update
     end
 
-
     def current_user_by_unsubscription_token
       token = UnsubscriptionToken.find_by(token: params[:unsubscription_token])
       if token
@@ -46,10 +43,6 @@ module EmailCampaigns
 
     def pundit_user
       current_user_by_unsubscription_token
-    end
-
-    def secure_controller?
-      false
     end
 
     private