+++ front/app/modules/commercial/customizable_navbar/admin/components/messages.ts
@@ -1,0 +1,8 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  tabNavigation: {
+    id: 'app.modules.navbar.admin.components.tabNavigation',
+    defaultMessage: 'Navigation',
+  },
+});